+++ src/embeddingLib/include/gradientOptimizer/SimpleOptimizer.hpp
@@ -4,16 +4,16 @@
 
 class SimpleOptimizer : public Optimizer {
    public:
-    SimpleOptimizer(int dimension, int numNodes, double maxDisplacement);
+    SimpleOptimizer(int dimension, int numNodes, float maxDisplacement);
     ~SimpleOptimizer();
 
-    void update(VecList& parameters, const VecList& gradients, double learningRate) override;
+    void update(VecList& parameters, const VecList& gradients, float learningRate) override;
     void reset() override;
 
    private:
     int dimension;
     int numNodes;
-    double maxDisplacement;
+    float maxDisplacement;
 
     VecList tmpGradient;
 };
