+++ src/embeddingLib/src/gradientOptimizer/AdamOptimizer.cpp
@@ -1,6 +1,6 @@
 #include "AdamOptimizer.hpp"
 
-AdamOptimizer::AdamOptimizer(int dimension, int numNodes, double beta1, double beta2, double epsilon)
+AdamOptimizer::AdamOptimizer(int dimension, int numNodes, float beta1, float beta2, float epsilon)
     : dimension(dimension),
       numNodes(numNodes),
       beta1(beta1),
@@ -12,7 +12,7 @@ AdamOptimizer::AdamOptimizer(int dimension, int numNodes, double beta1, double b
 
 AdamOptimizer::~AdamOptimizer() {}
 
-void AdamOptimizer::update(VecList& parameters, const VecList& gradients, double learningRate) {
+void AdamOptimizer::update(VecList& parameters, const VecList& gradients, float learningRate) {
     ASSERT(parameters.size() == numNodes, "Number of nodes in parameters does not match numNodes");
     ASSERT(gradients.size() == numNodes, "Number of nodes in gradients does not match numNodes");
 
@@ -22,8 +22,8 @@ void AdamOptimizer::update(VecList& parameters, const VecList& gradients, double
         for (int i = 0; i < dimension; i++) {
             m[n][i] = beta1 * m[n][i] + (1.0 - beta1) * gradients[n][i];
             v[n][i] = beta2 * v[n][i] + (1.0 - beta2) * gradients[n][i] * gradients[n][i];
-            double mHat = m[n][i] / (1.0 - pow(beta1, t));
-            double vHat = v[n][i] / (1.0 - pow(beta2, t));
+            float mHat = m[n][i] / (1.0 - pow(beta1, t));
+            float vHat = v[n][i] / (1.0 - pow(beta2, t));
             parameters[n][i] += learningRate * mHat / (sqrt(vHat) + epsilon);
         }
     }
