+++ src/graphLib/src/graph/GeometricGraphSampler.cpp
@@ -8,22 +8,22 @@
 #include "Rand.hpp"
 
 GraphCoordinatesPair GeometricGraphSampler::generateRandomGraphWithCoordinates(int n) {
-    double gridSize = std::sqrt(n);
-    double radius = std::sqrt(20.0 / M_PI);
+    float gridSize = std::sqrt(n);
+    float radius = std::sqrt(20.0 / M_PI);
     return generateRandomGraph(n, gridSize, radius);
 }
 
 Graph GeometricGraphSampler::generateRandomGraph(int n) { return generateRandomGraphWithCoordinates(n).first; }
 
-GraphCoordinatesPair GeometricGraphSampler::generateRandomGraph(int n, double gridSize, double radius) {
+GraphCoordinatesPair GeometricGraphSampler::generateRandomGraph(int n, float gridSize, float radius) {
     LOG_INFO("Constructing random graph...");
 
     // sample random coordinates
-    std::vector<std::vector<double>> coords(n);
+    std::vector<std::vector<float>> coords(n);
     for (int i = 0; i < n; i++) {
-        coords[i] = std::vector<double>(2);
-        coords[i][0] = Rand::randomDouble(0, gridSize);
-        coords[i][1] = Rand::randomDouble(0, gridSize);
+        coords[i] = std::vector<float>(2);
+        coords[i][0] = Rand::randomFloat(0, gridSize);
+        coords[i][1] = Rand::randomFloat(0, gridSize);
     }
 
     // build the graph (quadratic)
@@ -34,7 +34,7 @@ GraphCoordinatesPair GeometricGraphSampler::generateRandomGraph(int n, double gr
         }
         // check all higher nodes
         for (int u = v + 1; u < n; u++) {
-            double distance = std::sqrt((coords[u][0] - coords[v][0]) * (coords[u][0] - coords[v][0]) +
+            float distance = std::sqrt((coords[u][0] - coords[v][0]) * (coords[u][0] - coords[v][0]) +
                                         (coords[u][1] - coords[v][1]) * (coords[u][1] - coords[v][1]));
             if (distance < radius) {
                 graphMap[v].insert(u);
@@ -52,7 +52,7 @@ GraphCoordinatesPair GeometricGraphSampler::generateRandomGraph(int n, double gr
 
 GraphCoordinatesPair GeometricGraphSampler::findLargesConnectedComponent(GraphCoordinatesPair& graphCoords) {
     Graph unconnected = graphCoords.first;
-    std::vector<std::vector<double>> unconnectedCoords = graphCoords.second;
+    std::vector<std::vector<float>> unconnectedCoords = graphCoords.second;
     auto cc = GraphAlgo::calculateComponentId(unconnected);
 
     std::vector<int> connectedComponent = cc.first;
@@ -70,7 +70,7 @@ GraphCoordinatesPair GeometricGraphSampler::findLargesConnectedComponent(GraphCo
 
     // add nodes and edges to new graph
     std::map<int, std::set<int>> graphMap;
-    std::vector<std::vector<double>> connectedCoords(largestSize);
+    std::vector<std::vector<float>> connectedCoords(largestSize);
     std::unordered_map<int, int> nodeIdMapping;
     int currIdCounter = 0;
 
@@ -91,7 +91,7 @@ GraphCoordinatesPair GeometricGraphSampler::findLargesConnectedComponent(GraphCo
             }
 
             // write the coordinate to new vector
-            std::vector<double> tmp = unconnectedCoords[v];
+            std::vector<float> tmp = unconnectedCoords[v];
             connectedCoords[nodeIdMapping[v]] = tmp;
         }
     }
